add .cc files for writers.
