+++ src/io/xml_writer.h
@@ -15,109 +15,30 @@ You should have received a copy of the GNU Lesser General Public License along w
 #ifndef SRC_IO_XML_WRITER_H_
 #define SRC_IO_XML_WRITER_H_
 
-#include <any>
-#include <string>
 #include <vector>
 
 #include "pugixml.hpp"
 
-#include "any_casts.h"
-#include "b_spline.h"
-#include "nurbs.h"
-#include "xml_writer_utils.h"
+#include <any>
 
 namespace io {
 class XMLWriter {
  public:
   XMLWriter() = default;
 
-  void WriteFile(const std::vector<std::any> &splines, const char *filename) {
-    pugi::xml_document doc;
-    pugi::xml_node spline_list = doc.append_child("SplineList");
-    spline_list.append_attribute("NumberOfSplines") = std::to_string(splines.size()).c_str();
-    for (const auto &spline : splines) {
-      AddSpline(&spline_list, spline);
-    }
-    doc.save_file(filename, "  ", pugi::format_indent_attributes, pugi::encoding_utf8);
-  }
+  void WriteFile(const std::vector<std::any> &splines, const char *filename);
 
  private:
-  void AddSpline(pugi::xml_node *spline_list, const std::any &spline) {
-    int spline_dimension = util::AnyCasts::GetSplineDimension(spline);
-    if (spline_dimension == 1) {
-      Add1DSpline(spline_list, spline);
-    } else if (spline_dimension == 2) {
-      Add2DSpline(spline_list, spline);
-    } else if (spline_dimension == 3) {
-      Add3DSpline(spline_list, spline);
-    } else if (spline_dimension == 4) {
-      Add4DSpline(spline_list, spline);
-    }
-  }
+  void AddSpline(pugi::xml_node *spline_list, const std::any &spline);
 
-  void Add1DSpline(pugi::xml_node *spline_list, const std::any &spline) {
-    std::shared_ptr<spl::Spline<1>> spline_ptr = util::AnyCasts::GetSpline<1>(spline);
-    pugi::xml_node spline_node = spline_list->append_child("SplineEntry");
-    AddSplineAttributes(&spline_node, 1, spline_ptr->GetPointDim(), spline_ptr->GetNumberOfControlPoints());
-    AddControlPointVarNames(&spline_node, spline_ptr->GetPointDim());
-    io::XMLWriterUtils<1>::AddControlPointVars(&spline_node, spline_ptr);
-    if (util::AnyCasts::IsRational<1>(spline)) io::XMLWriterUtils<1>::AddWeights(&spline_node, spline);
-    io::XMLWriterUtils<1>::AddDegrees(&spline_node, spline_ptr);
-    io::XMLWriterUtils<1>::AddKnotVectors(&spline_node, spline_ptr);
-  }
+  void Add1DSpline(pugi::xml_node *spline_list, const std::any &spline);
+  void Add2DSpline(pugi::xml_node *spline_list, const std::any &spline);
+  void Add3DSpline(pugi::xml_node *spline_list, const std::any &spline);
+  void Add4DSpline(pugi::xml_node *spline_list, const std::any &spline);
 
-  void Add2DSpline(pugi::xml_node *spline_list, const std::any &spline) {
-    std::shared_ptr<spl::Spline<2>> spline_ptr = util::AnyCasts::GetSpline<2>(spline);
-    pugi::xml_node spline_node = spline_list->append_child("SplineEntry");
-    AddSplineAttributes(&spline_node, 2, spline_ptr->GetPointDim(), spline_ptr->GetNumberOfControlPoints());
-    AddControlPointVarNames(&spline_node, spline_ptr->GetPointDim());
-    io::XMLWriterUtils<2>::AddControlPointVars(&spline_node, spline_ptr);
-    if (util::AnyCasts::IsRational<2>(spline)) io::XMLWriterUtils<2>::AddWeights(&spline_node, spline);
-    io::XMLWriterUtils<2>::AddDegrees(&spline_node, spline_ptr);
-    io::XMLWriterUtils<2>::AddKnotVectors(&spline_node, spline_ptr);
-  }
+  void AddSplineAttributes(pugi::xml_node *spline_node, int spline_dimension, int space_dimension, int control_points);
 
-  void Add3DSpline(pugi::xml_node *spline_list, const std::any &spline) {
-    std::shared_ptr<spl::Spline<3>> spline_ptr = util::AnyCasts::GetSpline<3>(spline);
-    pugi::xml_node spline_node = spline_list->append_child("SplineEntry");
-    AddSplineAttributes(&spline_node, 3, spline_ptr->GetPointDim(), spline_ptr->GetNumberOfControlPoints());
-    AddControlPointVarNames(&spline_node, spline_ptr->GetPointDim());
-    io::XMLWriterUtils<3>::AddControlPointVars(&spline_node, spline_ptr);
-    if (util::AnyCasts::IsRational<3>(spline)) io::XMLWriterUtils<3>::AddWeights(&spline_node, spline);
-    io::XMLWriterUtils<3>::AddDegrees(&spline_node, spline_ptr);
-    io::XMLWriterUtils<3>::AddKnotVectors(&spline_node, spline_ptr);
-  }
-
-  void Add4DSpline(pugi::xml_node *spline_list, const std::any &spline) {
-    std::shared_ptr<spl::Spline<4>> spline_ptr = util::AnyCasts::GetSpline<4>(spline);
-    pugi::xml_node spline_node = spline_list->append_child("SplineEntry");
-    AddSplineAttributes(&spline_node, 4, spline_ptr->GetPointDim(), spline_ptr->GetNumberOfControlPoints());
-    AddControlPointVarNames(&spline_node, spline_ptr->GetPointDim());
-    io::XMLWriterUtils<4>::AddControlPointVars(&spline_node, spline_ptr);
-    if (util::AnyCasts::IsRational<4>(spline)) io::XMLWriterUtils<4>::AddWeights(&spline_node, spline);
-    io::XMLWriterUtils<4>::AddDegrees(&spline_node, spline_ptr);
-    io::XMLWriterUtils<4>::AddKnotVectors(&spline_node, spline_ptr);
-  }
-
-  void AddSplineAttributes(pugi::xml_node *spline_node, int spline_dimension, int space_dimension, int control_points) {
-    spline_node->append_attribute("splDim") = spline_dimension;
-    spline_node->append_attribute("spaceDim") = space_dimension;
-    spline_node->append_attribute("numOfCntrlPntVars") = space_dimension;
-    spline_node->append_attribute("numCntrlPnts") = control_points;
-  }
-
-  void AddControlPointVarNames(pugi::xml_node *spline_node, int space_dimension) {
-    pugi::xml_node names = spline_node->append_child("cntrlPntVarNames");
-    if (space_dimension == 1) {
-      names.append_child(pugi::node_pcdata).set_value("\n      x\n    ");
-    } else if (space_dimension == 2) {
-      names.append_child(pugi::node_pcdata).set_value("\n      x y\n    ");
-    } else if (space_dimension == 3) {
-      names.append_child(pugi::node_pcdata).set_value("\n      x y z\n    ");
-    } else if (space_dimension == 4) {
-      names.append_child(pugi::node_pcdata).set_value("\n      x y z t\n    ");
-    }
-  }
+  void AddControlPointVarNames(pugi::xml_node *spline_node, int space_dimension);
 };
 }  // namespace io
