+++ test/io/vtk_writer_test.cc
@@ -14,6 +14,7 @@ You should have received a copy of the GNU Lesser General Public License along w
 
 #include "gmock/gmock.h"
 
+#include "nurbs.h"
 #include "vtk_writer.h"
 
 using testing::Test;
